backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.26
numpy==2.2.1
beautifulsoup4==4.12.3
requests==2.33.0
httpx==0.28.1
pytest==9.0.3
pytest-asyncio==1.3.0
weasyprint==68.0
jinja2==3.1.5
supabase==2.15.0
python-dotenv==1.2.2
PyJWT==2.10.1
defusedxml==0.7.1
slowapi==0.1.9
Pillow==11.1.0
